# backend_laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TelEat API Documentation",
        "description": "Dokumentasi API untuk aplikasi pemesanan makanan TelEat (Laravel & React)",
        "contact": {
            "email": "admin@teleat.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Daftar user baru",
                "description": "Mendaftarkan akun baru dengan role ADMIN, MERCHANT, atau PELANGGAN.",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "username",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "budisantoso"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "budi@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "ADMIN",
                                            "PELANGGAN",
                                            "MERCHANT"
                                        ],
                                        "example": "PELANGGAN"
                                    },
                                    "foto_profil": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pendaftaran sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcde..."
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Login user",
                "description": "Masuk ke aplikasi menggunakan username dan password untuk mendapatkan token JWT.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "budisantoso"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcde..."
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Kredensial salah"
                    }
                }
            }
        },
        "/api/auth/google": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Login via Google",
                "description": "Masuk atau daftar otomatis menggunakan ID Token dari Google OAuth.",
                "operationId": "93ac7753bf4596edc151005072fa4b8f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_token"
                                ],
                                "properties": {
                                    "id_token": {
                                        "description": "Google ID Token",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcde..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Logout user",
                "description": "Menghapus access token saat ini.",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logout berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logout berhasil"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Data user saat ini",
                "description": "Mendapatkan informasi detail user beserta profilnya berdasarkan token Sanctum.",
                "operationId": "9d9bc556f927575aa492a88c28fe5908",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "budisantoso"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "budi@gmail.com"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "PELANGGAN"
                                        },
                                        "profile": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Gunakan token Sanctum (auth_token) hasil login untuk mengakses endpoint yang dilindungi",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}